test(po): split test_po_report_integration.py by feature (#691)
Extract the shared workdir/snapshot fixture into
_po_integration_helpers.py::BasePoIntegration and split the 568-LOC
test file into two feature-focused modules:

- test_po_report_integration.py: report-generation surface (collect,
  status, milestone-progress, stale-issues, generate-report, pr-flow)
- test_po_plan_integration.py: planning surface (parse-plan, adherence,
  trends, bootstrap-plan)

Each file now sits well under the 500-LOC oversized threshold. All 13
tests pass against edomata (no coverage change).

Closes #691
